refactor(schematics): 💡 ngx-migrate
