+++ libs/transloco-schematics/src/ngx-migrate/index.ts
@@ -1,25 +1,62 @@
-import {
-  Rule,
-  SchematicsException,
-  schematic,
-} from '@angular-devkit/schematics';
-import { from, map } from 'rxjs';
 
-import { run } from './ngx-translate-migration';
+import nodePath from 'node:path';
+
+import { Rule } from '@angular-devkit/schematics';
+import * as ora from 'ora';
+import { replaceInFile } from 'replace-in-file';
+
 import { SchemaOptions } from './schema';
+import { generateMatchers } from './migration-matchers';
 
 export default function (options: SchemaOptions): Rule {
-  // @ts-ignore
-  return (tree) => {
-    if (!options.path) {
-      throw new SchematicsException(`
-        Please specify the root source of your project.
-        (e.g. --path ./src/app)
-      `);
+  return async () => {
+    console.log('\x1b[4m%s\x1b[0m', '\nStarting migration script');
+    const dir = nodePath.resolve(process.cwd());
+
+    const path = nodePath.join(dir, options.path, '/**/*');
+    const { tsReplacements, htmlReplacements } = generateMatchers(path);
+
+    async function migrate(matchersArr, filesType) {
+      console.log(`\nMigrating ${filesType} files 📜`);
+      let spinner;
+      const isWindows = process.platform === 'win32';
+
+      for (let i = 0; i < matchersArr.length; i++) {
+        const { step, matchers } = matchersArr[i];
+        const msg = `Step ${i + 1}/${matchersArr.length}: Migrating ${step}`;
+        spinner = ora().start(msg);
+        const noFilesFound = [];
+        for (const matcher of matchers) {
+          try {
+            await replaceInFile({
+              ...matcher,
+              glob: {
+                windowsPathsNoEscape: isWindows,
+              },
+            });
+          } catch (e) {
+            if (e.message.includes('No files match the pattern')) {
+              noFilesFound.push(e.message);
+            } else {
+              throw e;
+            }
+          }
+        }
+        spinner.succeed(msg);
+        noFilesFound.forEach((pattern) =>
+          console.log('\x1b[33m%s\x1b[0m', `⚠️ ${pattern}`),
+        );
+      }
     }
 
-    return options.from === 'Angular'
-      ? schematic('ng-migrate', options)
-      : from(run(options.path)).pipe(map(() => tree));
+    return migrate(htmlReplacements, 'HTML')
+      .then(() => migrate(tsReplacements, 'TS'))
+      .then(() => {
+        console.log('\n              🌵 Done! 🌵');
+        console.log('Welcome to a better translation experience 🌐');
+        console.log(
+          '\nFor more information about this script please visit 👉 https://jsverse.github.io/transloco/docs/migration/ngx\n',
+        );
+      });
   };
 }
