+++ libs/transloco-schematics/src/ngx-migrate/schema.json
@@ -1,16 +1,9 @@
 {
   "$schema": "http://json-schema.org/schema",
   "$id": "SchematicsTranslocoMigrate",
-  "title": "Run a migration script from @ngx-translate.",
+  "title": "Migrate from @ngx-translate to Transloco",
   "type": "object",
   "properties": {
-    "from": {
-      "description": "Where is the migration from",
-      "x-prompt": "✍ Which library did you use?",
-      "enum": ["NgxTranslate", "Angular"],
-      "default": "NgxTranslate",
-      "type": "string"
-    },
     "path": {
       "description": "The path to the source root directory.",
       "x-prompt": "✍ Which folder will contain the translation files?",
