+++ libs/transloco-schematics/src/ngx-migrate/schema.ts
@@ -3,8 +3,4 @@ export interface SchemaOptions {
    * The path to the source root directory.
    */
   path: string;
-  /**
-   * Where is the migration from.
-   */
-  from: 'Angular' | 'NgxTranslate';
 }
